fix: Limit macOS fix to pyright-scip-side code changes (#178)
In PR #177, I added normalization logic inside setTrackedFiles.

However, it seems like the other file tracking already must be
taking care of passing normalized paths as necessary (hopefully?).

In order to minimize the diff size, I'm moving the normalization
to the call-site of setTrackedFiles inside the pyright-scip folder.
We want to do a merge of upstream changes sooner rather than later,
so it'd be good to reduce the diff.