fix: Set source maps support via API (#181)
`env` on different systems behaves differently. On macOS,
`env ABC XYZ` will invoke ABC with argument XYZ.
On some Linux variants, you need `env -S ABC XYZ` for that.
On Alpine, the `-S` flag is not supported, and there is seemingly
no way of passing arguments.

So this patch undoes the accidental breakage in #179
and uses the `getSourceMapsSupport` API.

Technically, both `process.setSourceMapsEnabled` and
`module.setSourceMapsSupport` are experimental functions,
so to be safe, we do a dynamic existence check before calling
them for portability across Node versions.