fix: Try locating packages using importlib (#196)
Changes the logic for listing packages to use importlib
which turns out to be much faster.

https://ampcode.com/threads/T-20d560ec-6811-4993-b63e-8935fa3d07f5#message-122-block-0

Adds an "integration" test around the Python script
so that we can be more confident that it works. The test
has both global and virtualenv packages, and both
are correctly located.

Also tested against `rich` and looks like the results
are unchanged.

The code could probably be cleaned up a bit, but I'm
prioritizing speed so that we can cut a release.